Add files via upload
This is a medical database. In this database, the user makes an account. This gives them the capability of determining what virus and/or disease they have based on symptom entry. Also, the user can check their history and see what symptoms, diseases, and viruses they had. They may also check when they had the symptoms, diseases, or viruses.